move travel screens to the right place and touchup some styles (#22)
