+++ go-corona-app/App.js
@@ -12,8 +12,7 @@ import { byPassLogin } from './constants/DevSettings';
 import RootView from './navigation/BottomTabNavigator';
 import useLinking from './navigation/useLinking';
 import LoginScreen from './screens/LoginScreen';
-import CheckupQuestionnaire from "./screens/checkup/CheckupScreen";
-import TravelQuestionnaire from "./screens/crosscheck/travel/Questionnaire";
+import TravelQuestionnaire from "./screens/crosscheck/travel/TravelScreen";
 import IntersectionCalculator from './screens/crosscheck/intersection/index';
 import { Provider as PaperProvider } from 'react-native-paper';
 
@@ -64,7 +63,6 @@ export default function App(props) {
           <AppStack.Navigator initialRouteName={byPassLogin ? "Root" : "Login"}>
             <AppStack.Screen name="Login" component={LoginScreen} />
             <AppStack.Screen name="Root" component={RootView} />
-            <AppStack.Screen name="CheckupQuestionnaire" options={{ title: "Checkup Questionnaire" }} component={CheckupQuestionnaire} />
             <AppStack.Screen name="TravelQuestionnaire" options={{ title: "Travel Questionnaire"}} component={TravelQuestionnaire} />
             <AppStack.Screen name="IntersectionCalculator" options={{ title: "Intersection Calculator"}} component={IntersectionCalculator} />
           </AppStack.Navigator>
